requirements.txt
torch
img2vec-pytorch
Pillow
scikit-learn
